+++ src/pages/basic-data/BasicDataPage.tsx
@@ -33,7 +33,7 @@ import { PersonRegisterPage } from '../basic_data/institution_admin/person_regis
 import { PublisherClaimsSettings } from '../editor/PublisherClaimsSettings';
 import { SerialPublicationClaimsSettings } from '../editor/SerialPublicationClaimsSettings';
 import NotFound from '../errorpages/NotFound';
-import { NviAdminNavigationAccordion } from './components/NviAdminNavigationAccordion';
+import { NviAdminNavigationAccordion } from './_components/NviAdminNavigationAccordion';
 import { NviAdminPublicationPointsPage } from './nvi/publication-points/NviAdminPublicationPointsPage';
 import { NviAdminReportingStatusPage } from './nvi/status/NviAdminReportingStatusPage';
