+++ src/pages/basic-data/nvi/_styles/nvi-admin-table-styles.ts
@@ -1,0 +1,10 @@
+import { styled } from '@mui/material';
+import { CenteredTableCell } from '../../../../components/tables/table-styles';
+
+export const CenteredPercentageControlledCell = styled(CenteredTableCell)({
+  minWidth: '9rem',
+});
+
+export const CenteredContactInformationCell = styled(CenteredTableCell)({
+  minWidth: '13rem',
+});
