--- src/pages/basic-data/nvi/components/styles/nvi-admin-table-styles.ts
@@ -1,10 +1,0 @@
-import { styled } from '@mui/material';
-import { CenteredTableCell } from '../../../../../components/tables/table-styles';
-
-export const CenteredPercentageControlledCell = styled(CenteredTableCell)({
-  minWidth: '9rem',
-});
-
-export const CenteredContactInformationCell = styled(CenteredTableCell)({
-  minWidth: '13rem',
-});
