+++ src/pages/basic-data/nvi/status/NviAdminReportingStatusPage.tsx
@@ -16,19 +16,16 @@ import { useFetchCustomerMap } from '../../../../api/hooks/useFetchCustomerMap';
 import { InstitutionContactInformationDialog } from '../../../../components/dialogs/institution-contact-information/InstitutionContactInformationDialog';
 import { NviPageLayout } from '../../../../components/page-layouts/NviPageLayout';
 import { TableSkeleton } from '../../../../components/skeletons/TableSkeleton';
-import { NviAdminSortSelectorType } from '../components/utils/nvi-admin-sort-types';
+import { NviAdminSortSelectorType } from '../_utils/nvi-admin-sort-types';
 import { VerticalBox } from '../../../../components/styled/Wrappers';
 import { CenteredTableCell } from '../../../../components/tables/table-styles';
 import { InstitutionReport } from '../../../../types/nvi.types';
 import { useNviCandidatesParams } from '../../../../utils/hooks/useNviCandidatesParams';
-import { useInstitutionReportsFilteredAndSortedByUrl } from '../components/hooks/useInstitutionReportsFilteredAndSortedByUrl';
-import { NviAdminTableSortSelector } from '../components/NviAdminTableSortSelector';
-import {
-  CenteredContactInformationCell,
-  CenteredPercentageControlledCell,
-} from '../components/styles/nvi-admin-table-styles';
-import { NviAdminReportingStatusRow } from './components/NviAdminReportingStatusRow';
-import { NviAdminReportingStatusTexts } from './components/NviAdminReportingStatusTexts';
+import { useInstitutionReportsFilteredAndSortedByUrl } from '../_hooks/useInstitutionReportsFilteredAndSortedByUrl';
+import { NviAdminTableSortSelector } from '../_components/NviAdminTableSortSelector';
+import { CenteredContactInformationCell, CenteredPercentageControlledCell } from '../_styles/nvi-admin-table-styles';
+import { NviAdminReportingStatusRow } from './_components/NviAdminReportingStatusRow';
+import { NviAdminReportingStatusTexts } from './_components/NviAdminReportingStatusTexts';
 
 export const NviAdminReportingStatusPage = () => {
   const { t } = useTranslation();
