+++ src/pages/tasks/nvi/publication-points/_components/NviPublicationPointsTexts.tsx
